requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
beautifulsoup4==4.12.2
crispy-bootstrap4==2023.1
crispy-bootstrap5==2023.10
dj-database-url==2.1.0
Django==4.2.7
django-bootstrap-datepicker==1.5.0
django-bootstrap-datepicker-plus @ git+https://github.com/monim67/django-bootstrap-datepicker-plus.git@b871d3c6056145f0fcd9fd99aaa77058da2814a4
django-bootstrap4==23.2
django-bootstrap5==23.3
django-cors-headers==4.3.0
django-crispy-forms==2.1
django-multiselectfield==0.1.12
gunicorn==21.2.0
method-decorator==0.1.3
ngrok==0.12.0
packaging==23.2
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
